比較コマンドをcommに変更

+++ HatebuCroller.sh
@@ -21,26 +21,35 @@ function WebCroller()
 function HtmlAnalyze()
 {
 	# Html File Exist Check
-	test -e ${FILE_TEMP_NEW} || exit
-
-	# Rename New File to Old File
-	mv ${FILE_TEMP_NEW} ${FILE_TEMP_OLD}
+	test -e ${FILE_TEMP_NEW} && mv ${FILE_TEMP_NEW} ${FILE_TEMP_OLD}
 
 	# Convert to News Title and URL
 	grep 'class="hb-entry-link-container"' ${FILE} \
-	| awk -Fdata-entryrank=\" {'print $1}' \
-	| awk -Fclass=\"entry-link\" '{print $2 $1 }' \
-	| sed -E 's/[[:space:]]+title="//g' \
-	| sed -E 's/\"[[:space:]]+<h3.*href="/ /g' \
-	| sed -e 's/\" $//g' \
+	| grep 'class="hb-entry-link-container"' hatebu.html \
+	| awk -Ftitle=\" '{print $2 $1}' \
+	| sed -E 's/\"[[:space:]]data-entryrank=.+<a[[:space:]]href=\"/ /g' \
+	| sed -E 's/\"[[:space:]]class=\".*//g' \
+	| sort \
 	> ${FILE_TEMP_NEW}
 }
 
 # Hatebu News Page Diff Check
 function HtmlDiffCheck()
 {
+	# Old File Check
+	test -e ${FILE_TEMP_OLD} || return
+
 	# Diff Check
-	diff ${FILE_TEMP_NEW} ${FILE_TEMP_OLD} | grep -e ^\< | sed -E 's/<[[:space:]]//g'
+	comm ${FILE_TEMP_NEW} ${FILE_TEMP_OLD} | grep -Ev ^[[:space:]]+ > ${FILE_TEMP_TMP} && TweetHatebu
+}
+
+# Twitter
+function TweetHatebu()
+{
+	while read NEWS_MSG; do
+		echo ${NEWS_MSG}{$NEWS_MSG}
+		#yes | tw ${NEWS_MSG}{$NEWS_MSG}
+	done < ${FILE_TEMP_TMP}
 }
 
 ##################
@@ -52,11 +61,13 @@ PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin
 # Target URL.
 URL=(
 "http://b.hatena.ne.jp/hotentry/it"
-"http://b.hatena.ne.jp/hotentry/general"
 )
 
 # Hatebu Html File.
 FILE='/tmp/hatebu.html'
+
+# Bot Message
+MSG_BOT='[自動ツイート]'
 
 # Analyze Function
 for (( count=0; count<${#URL[*]}; count++ ))
@@ -64,6 +75,7 @@ do
 	# Set Temp File Name.
 	FILE_TEMP_NEW=`echo /tmp/hatebu-``echo ${URL[$count]} | awk -F\/ '{print $5}'``echo .html.new`
 	FILE_TEMP_OLD=`echo /tmp/hatebu-``echo ${URL[$count]} | awk -F\/ '{print $5}'``echo .html.old`
+	FILE_TEMP_TMP=`echo /tmp/hatebu-``echo ${URL[$count]} | awk -F\/ '{print $5}'``echo .html.tmp`
 
 	# WebCroll.
 	WebCroller
